requirements.txt
# --- Core app dependencies ---
streamlit
ultralytics
opencv-python-headless
Pillow
matplotlib
numpy
requests
scikit-image
imagehash

# --- Machine Learning & Vector Search ---
torch
torchvision
faiss-cpu
open-clip-torch
transformers

